# driver-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/drivers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drivers"
                ],
                "summary": "List Drivers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.DriverListItem"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Yeni driver oluşturur",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drivers"
                ],
                "summary": "Create Driver",
                "parameters": [
                    {
                        "description": "Driver",
                        "name": "driver",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateDriverRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateDriverResponse"
                        }
                    }
                }
            }
        },
        "/drivers/nearby": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drivers"
                ],
                "summary": "Get Nearby Drivers",
                "parameters": [
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Longitude",
                        "name": "lon",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Taxi type",
                        "name": "taksiType",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.NearbyDriverResponse"
                            }
                        }
                    }
                }
            }
        },
        "/drivers/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drivers"
                ],
                "summary": "Update Driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driver ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Driver",
                        "name": "driver",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateDriverRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateDriverResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.CreateDriverRequest": {
            "type": "object",
            "properties": {
                "carBrand": {
                    "type": "string"
                },
                "carModel": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "plate": {
                    "type": "string"
                },
                "taxiType": {
                    "type": "string"
                }
            }
        },
        "dtos.CreateDriverResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dtos.DriverListItem": {
            "type": "object",
            "properties": {
                "carBrand": {
                    "type": "string"
                },
                "carModel": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "plate": {
                    "type": "string"
                },
                "taxiType": {
                    "type": "string"
                }
            }
        },
        "dtos.NearbyDriverResponse": {
            "type": "object",
            "properties": {
                "distanceKm": {
                    "type": "number"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "plate": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateDriverRequest": {
            "type": "object",
            "properties": {
                "carBrand": {
                    "type": "string"
                },
                "carModel": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "plate": {
                    "type": "string"
                },
                "taxiType": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateDriverResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}