# driver-service/dtos/nearby_driver_response.go
// dtos/nearby_driver_response.go
package dtos

type NearbyDriverResponse struct {
	FirstName  string  `json:"firstName"`
	LastName   string  `json:"lastName"`
	Plate      string  `json:"plate"`
	DistanceKm float64 `json:"distanceKm"`
}